Refactoring Merge two filters into one
